src: investigate if further refactoring of node_v8_platform-inl.h is of use
